Update File Operations feature for more performance.

+++ FileOperations.pde
@@ -12,10 +12,20 @@ final static boolean PRINT_FILE_OPERATIONS_FREE_EVENTS_DBG = false;
 //final static boolean PRINT_FILE_OPERATIONS_FREE_EVENTS_ERR = true;
 final static boolean PRINT_FILE_OPERATIONS_FREE_EVENTS_ERR = false;
 
+//final static boolean PRINT_FILE_OPERATIONS_FREE_ALWAYS_DBG = true;
+final static boolean PRINT_FILE_OPERATIONS_FREE_ALWAYS_DBG = false;
+//final static boolean PRINT_FILE_OPERATIONS_FREE_ALWAYS_ERR = true;
+final static boolean PRINT_FILE_OPERATIONS_FREE_ALWAYS_ERR = false;
+
 //final static boolean PRINT_FILE_OPERATIONS_SAVE_EVENTS_DBG = true;
 final static boolean PRINT_FILE_OPERATIONS_SAVE_EVENTS_DBG = false;
 //final static boolean PRINT_FILE_OPERATIONS_SAVE_EVENTS_ERR = true;
 final static boolean PRINT_FILE_OPERATIONS_SAVE_EVENTS_ERR = false;
+
+//final static boolean PRINT_FILE_OPERATIONS_SAVE_ALWAYS_DBG = true;
+final static boolean PRINT_FILE_OPERATIONS_SAVE_ALWAYS_DBG = false;
+//final static boolean PRINT_FILE_OPERATIONS_SAVE_ALWAYS_ERR = true;
+final static boolean PRINT_FILE_OPERATIONS_SAVE_ALWAYS_ERR = false;
 
 final static long FILE_OPERATIONS_FREE_LIMIT = 1L*1024L*1024L*1024L; // 1GB
 
@@ -105,6 +115,8 @@ void File_Operations_free_events()
 {
   String events_dir_full_name = sketchPath("events\\");
   File events_dir_handle = new File(events_dir_full_name);
+  Path events_dir_full_path = Paths.get(events_dir_full_name);
+  DirectoryStream<Path> events_dirs_list;
 
   delay(FRAME_TIME * 1 / 4);
   do
@@ -128,16 +140,22 @@ void File_Operations_free_events()
     if (free_space > FILE_OPERATIONS_FREE_LIMIT) continue;
     if (PRINT_FILE_OPERATIONS_ALL_DBG || PRINT_FILE_OPERATIONS_FREE_EVENTS_DBG) println("File_Operations_free_events():have not enough free_space="+free_space+" at "+events_dir_full_name);
 
-    // get files list.
-    String[] events_dirs_list = events_dir_handle.list();
-    if (PRINT_FILE_OPERATIONS_ALL_DBG || PRINT_FILE_OPERATIONS_FREE_EVENTS_DBG) println("File_Operations_free_events():events_dirs_list="+events_dirs_list);
-    if (events_dirs_list == null) continue;
-    if (PRINT_FILE_OPERATIONS_ALL_DBG || PRINT_FILE_OPERATIONS_FREE_EVENTS_DBG) println("File_Operations_free_events():events_dirs_list.length="+events_dirs_list.length);
-    //if (events_dirs_list.length <= 10) continue;
+    // Get dirs list.
+    try
+    {
+      events_dirs_list = Files.newDirectoryStream(events_dir_full_path);
+    }
+    catch (IOException e) {
+      // An I/O problem has occurred
+      if (PRINT_FILE_OPERATIONS_ALL_ERR || PRINT_FILE_OPERATIONS_FREE_EVENTS_ERR) println("File_Operations_free_events():Files.newDirectoryStream err!"+"\n\t"+e.toString());
+      continue;
+    }
 
-    //int delete_start_millis = millis();
-    for (String events_subdir_name:events_dirs_list) {
+    for (Path events_subdir_path:events_dirs_list)
+    {
+      String events_subdir_name = events_subdir_path.getFileName().toString();
       if (PRINT_FILE_OPERATIONS_ALL_DBG || PRINT_FILE_OPERATIONS_FREE_EVENTS_DBG) println("File_Operations_free_events():events_subdir_name="+events_subdir_name);
+
       File events_subdir_handle = new File(events_dir_full_name+events_subdir_name+"\\");
       if (!events_subdir_handle.isDirectory()) {
         if (PRINT_FILE_OPERATIONS_ALL_ERR || PRINT_FILE_OPERATIONS_FREE_EVENTS_ERR) println("File_Operations_free_events():not directory! "+events_dir_full_name+events_subdir_name+"\\");
@@ -145,17 +163,26 @@ void File_Operations_free_events()
         delay(1);
         continue;
       }
-      String[] events_subdir_files_list = events_subdir_handle.list();
-      if (PRINT_FILE_OPERATIONS_ALL_DBG || PRINT_FILE_OPERATIONS_FREE_EVENTS_DBG) println("File_Operations_free_events():events_subdir_files_list="+events_subdir_files_list);
-      if (events_subdir_files_list == null) {
+
+      // Get files list on subdir.
+      DirectoryStream<Path> events_subdir_files_list;
+      try
+      {
+        events_subdir_files_list = Files.newDirectoryStream(Paths.get(events_dir_full_name+events_subdir_name+"\\"));
+      }
+      catch (IOException e) {
+        // An I/O problem has occurred
+        if (PRINT_FILE_OPERATIONS_ALL_ERR || PRINT_FILE_OPERATIONS_FREE_EVENTS_ERR) println("File_Operations_free_events():Files.newDirectoryStream err!"+"\n\t"+e.toString());
         if (PRINT_FILE_OPERATIONS_ALL_DBG || PRINT_FILE_OPERATIONS_FREE_EVENTS_DBG) println("File_Operations_free_events():no files on sub directory! "+events_dir_full_name+events_subdir_name+"\\");
         events_subdir_handle.delete();
         delay(1);
         continue;
       }
-      if (PRINT_FILE_OPERATIONS_ALL_DBG || PRINT_FILE_OPERATIONS_FREE_EVENTS_DBG) println("File_Operations_free_events():events_subdir_files_list.length="+events_subdir_files_list.length);
-      for (String events_file_name:events_subdir_files_list) {
-        //println("File_Operations_free_events():events_file_name="+events_file_name);
+
+      for (Path events_file_path:events_subdir_files_list)
+      {
+        String events_file_name = events_file_path.getFileName().toString();
+        //if (PRINT_FILE_OPERATIONS_ALL_DBG || PRINT_FILE_OPERATIONS_FREE_EVENTS_DBG) println("File_Operations_free_events():events_file_name="+events_file_name);
         File events_file_handle = new File(events_dir_full_name+events_subdir_name+"\\"+events_file_name);
         if (!events_file_handle.isFile()) {
           if (PRINT_FILE_OPERATIONS_ALL_ERR || PRINT_FILE_OPERATIONS_FREE_EVENTS_ERR) println("File_Operations_free_events():not file! "+events_dir_full_name+events_subdir_name+"\\"+events_file_name);
@@ -164,6 +191,7 @@ void File_Operations_free_events()
         events_file_handle.delete();
         delay(1);
       }
+
       events_subdir_handle.delete();
       delay(1);
       if (PRINT_FILE_OPERATIONS_ALL_DBG || PRINT_FILE_OPERATIONS_FREE_EVENTS_DBG) println("File_Operations_free_events():delete done! "+events_subdir_name);
@@ -174,26 +202,44 @@ void File_Operations_free_events()
 
 void File_Operations_free_always()
 {
+  String always_dir_full_name = sketchPath("always\\");
+  File always_dir_handle = new File(always_dir_full_name);
+  Path always_dir_full_path = Paths.get(always_dir_full_name);
+  DirectoryStream<Path> always_files_list;
+  
   delay(FRAME_TIME * 2 / 4);
   do
   {
     delay(FRAME_TIME);
     if (File_Operations_free_threads_pause) continue;
 
-    if (PRINT_FILE_OPERATIONS_ALL_DBG) println("File_Operations_free_always():Run"+millis());
+    if (PRINT_FILE_OPERATIONS_ALL_DBG || PRINT_FILE_OPERATIONS_FREE_ALWAYS_DBG) println("File_Operations_free_always():Run"+millis());
     //println("File_Operations_free_always():Run"+millis());
+
+    if (!always_dir_handle.isDirectory()) {
+      if (PRINT_FILE_OPERATIONS_ALL_DBG || PRINT_FILE_OPERATIONS_FREE_ALWAYS_DBG) println("File_Operations_free_always():always directory is not ready!");
+      continue;
+    }
+
+    // Get files list to decide delete file.
+    try
+    {
+      always_files_list = Files.newDirectoryStream(always_dir_full_path);
+    }
+    catch (IOException e) {
+      // An I/O problem has occurred
+      if (PRINT_FILE_OPERATIONS_ALL_ERR || PRINT_FILE_OPERATIONS_FREE_ALWAYS_ERR) println("File_Operations_free_always():Files.newDirectoryStream err!"+"\n\t"+e.toString());
+      continue;
+    }
 
     long date_time = new Date().getTime();
     //Dbg_Time_logs_handle.add("Date().getTime()");
-    String always_dir_full_name = sketchPath("always\\");
-    File always_dir_handle = new File(always_dir_full_name);
-    //println("date_time="+date_time+",millis()="+millis());
-    if (!always_dir_handle.isDirectory()) continue;
 
-    // get files list to decide delete file.
-    String[] always_files_list = always_dir_handle.list();
-    if (always_files_list == null) continue;
-    for (String always_file_name:always_files_list) {
+    for (Path always_file_path:always_files_list)
+    {
+      String always_file_name = always_file_path.getFileName().toString();
+      //println("File_Operations_free_always():always_file_name="+always_file_name);
+
       long file_date_time;
       try {
         file_date_time = Long.parseLong(always_file_name.substring(2, always_file_name.length() - 4));
@@ -203,8 +249,13 @@ void File_Operations_free_always()
       }
       if (file_date_time > date_time - PS_DATA_SAVE_ALWAYS_DURATION*60*60*1000L) continue;
       //println(always_file_name+","+file_date_time);
+
       File always_file_handle;
       always_file_handle = new File(always_dir_full_name+always_file_name);
+      if (!always_file_handle.isFile()) {
+        if (PRINT_FILE_OPERATIONS_ALL_ERR || PRINT_FILE_OPERATIONS_FREE_ALWAYS_ERR) println("File_Operations_free_always():not file! "+always_dir_full_name+"\\"+always_file_name);
+        continue;
+      }
       always_file_handle.delete();
       delay(1);
     }
@@ -214,7 +265,7 @@ void File_Operations_free_always()
 void File_Operations_save_events()
 {
   String always_dir_full_name = sketchPath("always\\");
-  Path always_dir_path = Paths.get(always_dir_full_name);
+  Path always_dir_full_path = Paths.get(always_dir_full_name);
   DirectoryStream<Path>[] always_files_list = new DirectoryStream[PS_INSTANCE_MAX];
   Iterator<Path>[] always_files_list_iterator = new Iterator[PS_INSTANCE_MAX];
 
@@ -240,6 +291,10 @@ void File_Operations_save_events()
       {
         delay(FRAME_TIME);
       }
+      else
+      {
+        delay(1);
+      }
 
       //println("File_Operations_save_events():"+":File_Operations_save_events_state["+instance+"]="+File_Operations_save_events_state[instance]);
       switch (File_Operations_save_events_state[instance])
@@ -248,14 +303,14 @@ void File_Operations_save_events()
           // Set pause state of Disk Space free threads to save events files.
           File_Operations_free_threads_pause = true;
 
-          // get files list of instance.
+          // Get files list of instance.
           try
           {
-            always_files_list[instance] = Files.newDirectoryStream(always_dir_path, instance+"_*.dat");
+            always_files_list[instance] = Files.newDirectoryStream(always_dir_full_path, instance+"_*.dat");
           }
           catch (IOException e) {
             // An I/O problem has occurred
-            if (PRINT_FILE_OPERATIONS_ALL_ERR) println("File_Operations_save_events("+instance+"):Files.newDirectoryStream err!"+"\n\t"+e.toString());
+            if (PRINT_FILE_OPERATIONS_ALL_ERR || PRINT_FILE_OPERATIONS_SAVE_EVENTS_ERR) println("File_Operations_save_events("+instance+"):Files.newDirectoryStream err!"+"\n\t"+e.toString());
             // Reset pause state of Disk Space free threads to save events files.
             File_Operations_free_threads_pause = false;
             break;
@@ -265,16 +320,24 @@ void File_Operations_save_events()
           always_files_list_iterator[instance] = always_files_list[instance].iterator();
 
           File_Operations_save_events_state[instance] = File_Operations_save_events_state_enum.COPY_ALWAYS_TO_EVENTS;
+
           break;
 
         case COPY_ALWAYS_TO_EVENTS:
           int copy_start_millis = millis();
+          File always_file_handle;
           for (; always_files_list_iterator[instance].hasNext();)
           {
             Path always_file_path = always_files_list_iterator[instance].next();
 
             String always_file_name = always_file_path.getFileName().toString();
             //println("File_Operations_save_events("+instance+"):always_file_name="+always_file_name);
+
+            always_file_handle = new File(always_dir_full_name+always_file_name);
+            if (!always_file_handle.isFile()) {
+              if (PRINT_FILE_OPERATIONS_ALL_ERR || PRINT_FILE_OPERATIONS_SAVE_EVENTS_ERR) println("File_Operations_save_events():not file! "+always_dir_full_name+"\\"+always_file_name);
+              continue;
+            }
 
             // Get date time of always file.
             long always_file_date_time;
@@ -298,37 +361,48 @@ void File_Operations_save_events()
                   File_Operations_save_events_dir_full_name[instance] + always_file_name,
                   new CopyOption[] {StandardCopyOption.COPY_ATTRIBUTES}))
             {
-              if (PRINT_FILE_OPERATIONS_ALL_ERR) println("File_Operations_save_events():" + instance + ":copy_file() error!" + "\n\t" + always_dir_full_name + always_file_name + "->" + "\n\t" + File_Operations_save_events_dir_full_name[instance] + always_file_name + "\n\t" + copy_file_error);
+              if (PRINT_FILE_OPERATIONS_ALL_ERR || PRINT_FILE_OPERATIONS_SAVE_EVENTS_ERR) println("File_Operations_save_events():" + instance + ":copy_file() error!" + "\n\t" + always_dir_full_name + always_file_name + "->" + "\n\t" + File_Operations_save_events_dir_full_name[instance] + always_file_name + "\n\t" + copy_file_error);
               break;
             }
             //delay(1);
             // Check copy operation is too late by frame time.
-            if (get_millis_diff(copy_start_millis) >= FRAME_TIME) break;
+            if (get_millis_diff(copy_start_millis) > FRAME_TIME) break;
           }
-          if (!always_files_list_iterator[instance].hasNext())
+
+          // Check end of iterator.
+          if (always_files_list_iterator[instance].hasNext())
           {
-            try
-            {
-              always_files_list[instance].close();
-            }
-            catch (IOException e) {
-              // An I/O problem has occurred
-              if (PRINT_FILE_OPERATIONS_ALL_ERR) println("File_Operations_save_events("+instance+"):DirectoryStream close() err!"+"\n\t"+e.toString());
-            }
-            File_Operations_save_events_state[instance] = File_Operations_save_events_state_enum.WAIT_WRITE_EVENTS_DONE;
-            // Reset pause state of Disk Space free threads to save events files.
-            File_Operations_free_threads_pause = false;
+            break;
           }
+
+          // Reset list.
+          try
+          {
+            always_files_list[instance].close();
+          }
+          catch (IOException e) {
+            // An I/O problem has occurred
+            if (PRINT_FILE_OPERATIONS_ALL_ERR || PRINT_FILE_OPERATIONS_SAVE_EVENTS_ERR) println("File_Operations_save_events("+instance+"):DirectoryStream close() err!"+"\n\t"+e.toString());
+          }
+
+          File_Operations_save_events_state[instance] = File_Operations_save_events_state_enum.WAIT_WRITE_EVENTS_DONE;
+          // Reset pause state of Disk Space free threads to save events files.
+          File_Operations_free_threads_pause = false;
+
           break;
 
         case WAIT_WRITE_EVENTS_DONE:
+          // Check write done of save events.
           if (!File_Operations_save_events_write_events_done[instance]) break;
+
           File_Operations_save_events_state[instance] = File_Operations_save_events_state_enum.DONE;
+
           break;
 
         case DONE:
           File_Operations_save_events_done[instance] = true;
           File_Operations_save_events_state[instance] = File_Operations_save_events_state_enum.IDLE;
+
           break;
       }
     }
